requirements.txt
pyside6
sqlalchemy
python-dotenv
pywin32
openai
pypdf
python-docx